refactor: :truck: move Header folder to appropriate place
