+++ institutional-site/src/components/Layout/index.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from 'react'
-import Header from '../Header'
+import Header from './components/Header'
 
 type Layout = {
   children: ReactNode
